+++ src/main/java/es/in2/desmos/domain/exceptions/JWTClaimMissingException.java
@@ -1,0 +1,14 @@
+package es.in2.desmos.domain.exceptions;
+
+import java.io.Serial;
+
+public class JWTClaimMissingException extends RuntimeException{
+
+    @Serial
+    private static final long serialVersionUID = 1L;
+
+    public JWTClaimMissingException(String message) {
+        super(message);
+    }
+
+}