+++ src/main/java/es/in2/desmos/domain/exceptions/TokenFetchException.java
@@ -1,0 +1,13 @@
+package es.in2.desmos.domain.exceptions;
+
+import java.io.Serial;
+
+public class TokenFetchException extends RuntimeException {
+
+    @Serial
+    private static final long serialVersionUID = 1L;
+
+    public TokenFetchException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}