+++ src/main/java/es/in2/desmos/domain/exceptions/handler/GlobalExceptionHandler.java
@@ -3,8 +3,6 @@ package es.in2.desmos.domain.exceptions.handler;
 import es.in2.desmos.domain.exceptions.*;
 import es.in2.desmos.domain.models.GlobalErrorMessage;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.io.buffer.DataBufferLimitException;
-import org.springframework.core.io.buffer.DataBufferUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -14,7 +12,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.bind.support.WebExchangeBindException;
 import reactor.core.publisher.Mono;
 
-import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -156,5 +153,23 @@ public class GlobalExceptionHandler {
         log.error("InvalidTokenException: {}", invalidTokenException.getMessage());
         String path = String.valueOf(request.getPath());
         return Mono.just(GlobalErrorMessage.builder().title("InvalidTokenException").message(invalidTokenException.getMessage()).path(path).build());
+    }
+
+    @ExceptionHandler(EntitySyncException.class)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ResponseBody
+    public Mono<GlobalErrorMessage> handleEntitySyncException(EntitySyncException entitySyncException, ServerHttpRequest request) {
+        log.error("EntitySyncException: {}", entitySyncException.getMessage());
+        String path = String.valueOf(request.getPath());
+        return Mono.just(GlobalErrorMessage.builder().title("EntitySyncException").message(entitySyncException.getMessage()).path(path).build());
+    }
+
+    @ExceptionHandler(DiscoverySyncException.class)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ResponseBody
+    public Mono<GlobalErrorMessage> handleDiscoverySyncException(DiscoverySyncException discoverySyncException, ServerHttpRequest request) {
+        log.error("DiscoverySyncException: {}", discoverySyncException.getMessage());
+        String path = String.valueOf(request.getPath());
+        return Mono.just(GlobalErrorMessage.builder().title("DiscoverySyncException").message(discoverySyncException.getMessage()).path(path).build());
     }
 }