+++ src/main/java/es/in2/desmos/domain/models/OIDCAccessTokenResponse.java
@@ -1,0 +1,11 @@
+package es.in2.desmos.domain.models;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+
+@Builder
+public record OIDCAccessTokenResponse(
+        @JsonProperty("access_token") String accessToken,
+        @JsonProperty("token_type") String tokenType,
+        @JsonProperty("expires_in") String expiresIn
+) {}