+++ src/main/java/es/in2/desmos/infrastructure/configs/TrustFrameworkConfig.java
@@ -30,8 +30,6 @@ public class TrustFrameworkConfig {
 
     private final Scheduler scheduler = Schedulers.boundedElastic();
 
-    private final AtomicReference<HashMap<String, String>> publicKeysByUrlRef = new AtomicReference<>();
-
     private final AtomicReference<HashSet<String>> dltAddressesRef = new AtomicReference<>();
 
     @Bean
@@ -45,22 +43,10 @@ public class TrustFrameworkConfig {
 
         return getAccessNodesListContent()
                 .flatMap(yamlAccessNodesList -> {
-                    HashMap<String, String> publicKeysByUrl = deserializePublicKeysByUrl(yamlAccessNodesList);
-                    savePublicKeysByUrlRef(Mono.just(publicKeysByUrl));
-
                     HashSet<String> dltAddresses = deserializeDltAddress(yamlAccessNodesList);
                     saveDltAddressesRef(Mono.just(dltAddresses));
                     return Mono.empty();
                 });
-    }
-
-    public HashMap<String, String> getPublicKeysByUrl() {
-        var publicKeysByUrl = publicKeysByUrlRef.get();
-        if (publicKeysByUrl == null || publicKeysByUrl.isEmpty()) {
-            return null;
-        } else {
-            return publicKeysByUrl;
-        }
     }
 
     public HashSet<String> getDltAddresses() {
@@ -70,12 +56,6 @@ public class TrustFrameworkConfig {
         } else {
             return dltAddresses;
         }
-    }
-
-    private void savePublicKeysByUrlRef(Mono<HashMap<String, String>> publicKeysByUrl) {
-        publicKeysByUrl
-                .publishOn(scheduler)
-                .subscribe(publicKeysByUrlRef::set);
     }
 
     private void saveDltAddressesRef(Mono<HashSet<String>> dltAddresses) {
@@ -92,32 +72,10 @@ public class TrustFrameworkConfig {
                     .get()
                     .uri(trustedAccessNodesListUri)
                     .retrieve()
-                    .bodyToMono(String.class);
+                    .bodyToMono(String.class)
+                    .retry(3);
         } catch (URISyntaxException e) {
             return Mono.error(new RuntimeException(e));
-        }
-    }
-
-    private HashMap<String, String> deserializePublicKeysByUrl(String yamlContent) {
-        try {
-            JsonNode rootNode = yamlMapper.readTree(yamlContent);
-
-            HashMap<String, String> resultMap = new HashMap<>();
-
-            JsonNode organizations = rootNode.path("organizations");
-            if (organizations.isArray()) {
-                for (JsonNode organization : organizations) {
-                    String url = organization.path("url").asText();
-                    String publicKeyDecimalString = organization.path("publicKey").asText();
-                    String hexString = decimalToHex64(publicKeyDecimalString, 130);
-                    resultMap.put(url, hexString);
-                }
-            }
-
-            return resultMap;
-
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
         }
     }